+++ tests/test_db.py
@@ -1,8 +1,9 @@
 
 
 import pytest
-from mailadm.conn import DBError, TokenExhausted
-from mailadm.util import get_doveadm_pw
+
+from mailadm.conn import DBError, TokenExhausted, UserNotFound
+from mailadm.util import gen_password
 
 
 def test_token(tmpdir, make_db):
@@ -44,50 +45,31 @@ class TestTokenAccounts:
 
     def test_add_with_wrong_token(self, conn):
         now = 10000
-        addr = "tmp.123@example.org"
-        clear_pw, hash_pw = get_doveadm_pw()
+        addr = "tmp.123@x.testrun.org"
         with pytest.raises(DBError):
-            conn.add_user(addr=addr, hash_pw=hash_pw,
-                          date=now, ttl=60 * 60, token_name="112l3kj123123")
+            conn.add_user_db(addr=addr, date=now, ttl=60 * 60, token_name="112l3kj123123")
 
     def test_add_maxuse(self, conn):
         now = 10000
-        clear_pw, hash_pw = get_doveadm_pw()
+        password = gen_password()
         for i in range(self.MAXUSE):
-            addr = "tmp.{}@example.org".format(i)
-            conn.add_user(addr=addr, hash_pw=hash_pw, date=now, ttl=60 * 60, token_name="onehour")
+            addr = "tmp.{}@x.testrun.org".format(i)
+            conn.add_user_db(addr=addr, date=now, ttl=60 * 60, token_name="onehour")
 
+        token_info = conn.get_tokeninfo_by_name("onehour")
         with pytest.raises(TokenExhausted):
-            conn.add_user(addr="tmp.xx@example.org", hash_pw=hash_pw,
-                          date=now, ttl=60 * 60, token_name="onehour")
-
-    def test_homedirs(self, conn):
-        clear_pw, hash_pw = get_doveadm_pw()
-
-        conn.add_user(
-            addr="tmp.1@example.org", hash_pw=hash_pw,
-            date=10, ttl=60 * 60, token_name="onehour")
-        conn.add_user(
-            addr="tmp.2@example.org", hash_pw=hash_pw,
-            date=11, ttl=60 * 60, token_name="onehour")
-        known = set()
-        for user_info in conn.get_user_list():
-            if user_info.homedir in known:
-                pytest.fail("duplicate homedir" + str(user_info.homedir))
-            assert user_info.homedir.relative_to(conn.config.path_vmaildir)
-            known.add(user_info.homedir)
+            conn.add_email_account(token_info, addr="tmp.xx@x.testrun.org", password=password)
 
     def test_add_expire_del(self, conn):
         now = 10000
-        addr = "tmp.123@example.org"
-        addr2 = "tmp.456@example.org"
-        addr3 = "tmp.789@example.org"
-        clear_pw, hash_pw = get_doveadm_pw()
-        conn.add_user(addr=addr, hash_pw=hash_pw, date=now, ttl=60 * 60, token_name="onehour")
+        addr = "tmp.123@x.testrun.org"
+        addr2 = "tmp.456@x.testrun.org"
+        addr3 = "tmp.789@x.testrun.org"
+        conn.add_user_db(addr=addr, date=now, ttl=60 * 60, token_name="onehour")
         with pytest.raises(DBError):
-            conn.add_user(addr=addr, hash_pw=hash_pw, date=now, ttl=60 * 60, token_name="onehour")
-        conn.add_user(addr=addr2, hash_pw=hash_pw, date=now, ttl=30 * 60, token_name="onehour")
-        conn.add_user(addr=addr3, hash_pw=hash_pw, date=now, ttl=32 * 60, token_name="onehour")
+            conn.add_user_db(addr=addr, date=now, ttl=60 * 60, token_name="onehour")
+        conn.add_user_db(addr=addr2, date=now, ttl=30 * 60, token_name="onehour")
+        conn.add_user_db(addr=addr3, date=now, ttl=32 * 60, token_name="onehour")
         conn.commit()
         expired = conn.get_expired_users(sysdate=now + 31 * 60)
         assert len(expired) == 1
@@ -95,11 +77,11 @@ class TestTokenAccounts:
 
         users = conn.get_user_list()
         assert len(users) == 3
-        conn.del_user(addr2)
+        conn.del_user_db(addr2)
         conn.commit()
         assert len(conn.get_user_list()) == 2
         addrs = [u.addr for u in conn.get_user_list()]
         assert addrs == [addr, addr3]
-        with pytest.raises(DBError):
-            conn.del_user(addr2)
+        with pytest.raises(UserNotFound):
+            conn.del_user_db(addr2)
         assert conn.get_tokeninfo_by_name("onehour").usecount == 3