+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+Unreleased
+-------------
+
+- breaking change: mailadm now uses mailcow REST API for creating/manipulating e-mail accounts instead of fiddling with postfix/dovecot config
+- deprecate dovecot/postfix support
+- provider instructions for docker setup
 
 0.10.5
 -------------